Remove AngularJS
