+++ .prettierrc
@@ -4,8 +4,7 @@
   "arrowParens": "always",
   "overrides": [
     {
-      "files": "*.html",
-      "options": { "parser": "angular" }
+      "files": "*.html"
     }
   ]
 }
