+++ package.json
@@ -52,9 +52,6 @@
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
     "@sentry/cli": "^1.30.4",
-    "@types/angular": "^1.6.43",
-    "@types/angular-hotkeys": "^1.7.0",
-    "@types/angular-mocks": "^1.5.11",
     "@types/lodash": "^4.14.0",
     "@types/memoize-one": "^4.1.0",
     "@types/node": "^10.9.3",
@@ -69,7 +66,6 @@
     "autoprefixer": "^9.0.1",
     "awesome-typescript-loader": "^5.0.0",
     "babel-loader": "^8.0.2",
-    "babel-plugin-angularjs-annotate": "^0.10.0",
     "babel-plugin-lodash": "^3.3.2",
     "case-sensitive-paths-webpack-plugin": "^2.1.2",
     "clean-webpack-plugin": "^1.0.0",
@@ -127,13 +123,8 @@
     "@fortawesome/free-solid-svg-icons": "^5.3.1",
     "@fortawesome/react-fontawesome": "^0.1.3",
     "@sentry/browser": "^4.0.6",
-    "@uirouter/angularjs": "^1.0.15",
     "@uirouter/react": "^0.8.2",
     "@uirouter/visualizer": "^7.0.0",
-    "angular": "^1.6.9",
-    "angular-aria": "^1.6.9",
-    "angular-sanitize": "^1.6.9",
-    "angular2react": "^3.0.2",
     "body-scroll-lock": "^2.6.1",
     "browserslist": "^4.0.0",
     "bungie-api-ts": "^1.6.0",
@@ -152,8 +143,6 @@
     "memoize-one": "^5.0.0",
     "mobile-drag-drop": "^2.3.0-rc.1",
     "mousetrap": "^1.6.2",
-    "ng-i18next": "^1.0.5",
-    "ngimport": "^1.0.0",
     "papaparse": "^4.6.2",
     "popper.js": "^1.14.3",
     "react": "^16.3.1",
@@ -171,7 +160,6 @@
     "react-twitter-widgets": "^1.7.1",
     "react-view-pager": "^0.6.0",
     "react-with-gesture": "^4.0.2",
-    "react2angular": "^4.0.1",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0",
     "reselect": "^4.0.0",
