--- src/app/app.module.ts
@@ -1,5 +1,0 @@
-import { module } from 'angular';
-
-import { DimAppModule } from './dimApp.module';
-
-export const AppModule = module('app', [DimAppModule]).name;
