--- src/app/dimApp.config.ts
@@ -1,7 +1,0 @@
-import { ICompileProvider } from 'angular';
-
-export default function config($compileProvider: ICompileProvider) {
-  'ngInject';
-
-  $compileProvider.imgSrcSanitizationWhitelist(/^\s*(https?:|data:image\/)/);
-}
