+++ src/app/inventory/InventoryItem.tsx
@@ -5,8 +5,7 @@ import './InventoryItem.scss';
 import { TagValue, itemTags } from './dim-item-info';
 import getBadgeInfo from './get-badge-info';
 import BungieImage, { bungieBackgroundStyle } from '../dim-ui/BungieImage';
-import { percent } from './dimPercentWidth.directive';
-import { getColor } from '../shell/dimAngularFilters.filter';
+import { getColor, percent } from '../shell/filters';
 import { AppIcon, lockIcon, thumbsUpIcon } from '../shell/icons';
 import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 import { InventoryCuratedRoll } from '../curated-rolls/curatedRollService';
