+++ src/app/item-popup/ItemPopupBody.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { DimItem } from '../inventory/item-types';
 import { t } from 'i18next';
-import { percent } from '../inventory/dimPercentWidth.directive';
 import ItemOverview from './ItemDetails';
 import { ItemPopupExtraInfo } from './item-popup';
 import ItemActions from './ItemActions';
 import classNames from 'classnames';
 import ItemReviews from '../item-review/ItemReviews';
+import { percent } from '../shell/filters';
 
 export enum ItemPopupTab {
   Overview,
