+++ src/app/item-popup/ItemStats.tsx
@@ -3,8 +3,7 @@ import { DimItem, DimStat, D1Stat } from '../inventory/item-types';
 import classNames from 'classnames';
 import { t } from 'i18next';
 import './ItemStats.scss';
-import { percent } from '../inventory/dimPercentWidth.directive';
-import { getColor } from '../shell/dimAngularFilters.filter';
+import { getColor, percent } from '../shell/filters';
 import { AppIcon, helpIcon } from '../shell/icons';
 import ExternalLink from '../dim-ui/ExternalLink';
 import * as _ from 'lodash';
