+++ src/app/settings/settings.ts
@@ -1,5 +1,4 @@
 import i18next from 'i18next';
-import { $rootScope } from 'ngimport';
 import * as _ from 'lodash';
 import { SyncService } from '../storage/sync.service';
 import store from '../store/store';
@@ -44,18 +43,16 @@ export function initSettings() {
 
     const savedSettings = (data['settings-v1.0'] || {}) as Partial<Settings>;
 
-    $rootScope.$evalAsync(() => {
-      const languageChanged = savedSettings.language !== i18next.language;
-      store.dispatch(loaded(savedSettings));
-      const settings = store.getState().settings;
-      localStorage.setItem('dimLanguage', settings.language);
-      if (languageChanged) {
-        i18next.changeLanguage(settings.language);
-      }
+    const languageChanged = savedSettings.language !== i18next.language;
+    store.dispatch(loaded(savedSettings));
+    const settings = store.getState().settings;
+    localStorage.setItem('dimLanguage', settings.language);
+    if (languageChanged) {
+      i18next.changeLanguage(settings.language);
+    }
 
-      readyResolve();
-      // Start saving settings changes
-      unsubscribe = saveSettingsOnUpdate();
-    });
+    readyResolve();
+    // Start saving settings changes
+    unsubscribe = saveSettingsOnUpdate();
   });
 }
