+++ src/index.html
@@ -37,7 +37,6 @@
   </head>
 
   <body>
-    <div id="angular"></div>
     <div id="app">
       <div class="notloaded">
         DIM hasn't loaded. Something may be wrong with it, or with your browser (maybe you have a
