--- src/scss/_debug.scss
@@ -1,11 +1,0 @@
-.debugItem {
-  max-width: 80em;
-
-  .section {
-    margin: 2em 0;
-  }
-
-  pre {
-    font-size: 120%;
-  }
-}
