+++ src/scss/main.scss
@@ -8,8 +8,6 @@
 
 @import 'animations';
 @import 'style';
-@import 'debug';
-@import 'toast';
 
 // Load stylesheets from external dependencies from node_modules
 @import '~components-font-awesome/css/font-awesome.css';
